Added float to int convertor in screens calculation

+++ qdrawsettings.py
@@ -78,8 +78,10 @@ class QdrawSettings(QWidget):
     def center(self):
         screen = QDesktopWidget().screenGeometry()
         size = self.geometry()
-        self.move((screen.width() - size.width()) / 2,
-                  (screen.height() - size.height()) / 2)
+        self.move(
+            int((screen.width() - size.width()) / 2),
+            int((screen.height() - size.height()) / 2)
+        )
 
     def closeEvent(self, e):
         self.clear()
